added individual sounds to each button
Dr. Yu told us to use this.innerHTML, but that didn't initially work. Kudos to a classmate who figured it out, but then another classmate discovered that it still works upon changing it back to this.HTML.  I tried it and it's true!  Peculiar.  Spent two hours trying to figure out what was wrong with my code.