Final Solution
I was unable to solve the detecting keyboard presses, but I did take the time to understand it once studying and completing the solution.

In my attempt, I had unsuccessfully tried to create a function that would make a variable keyStroke = buttonInnerHTML if keyStroke === buttonInnerHTML. My thinking was that this would convert the keydown to a button click and therefore run the switch function whenever the right key was pressed, but it didn't work.

It makes so much sense to make a separate function and to call the function  after the event listeners detect either a button click or a key press.

One thing that I did not understand was the event.key, but Dr. Yu explained that it is the key value of the event that we want. Then I recalled how many values the keydown event has, and it made sense that we would only want the one that would produce a single letter.

So in a nutshell, this.innerHTML and event.key are the two very different ways of generating the same thing--a single letter--the key that gets passed into the switch function.  I GET IT!!!