forgot to note this unaddressed error
i noticed on my own that leaving the console.log calling buttonInnerHTML causes an uncaught reference error: buttonInnerHTML is not defined
    at makeSound (index.js:57:30)
    at HTMLDocument.<anonymous> (index.js:15:5)
makeSound @ index.js:57
(anonymous) @ index.js:15.

I was able to fix it by calling key instead.  and now if a key other than the cases is pressed, the value of the key is logged.  in the event that a different button is somehow clicked, that would be logged as well.  As Dr. Yu had explained earlier when we created the default, it is just good practice to set it up, even if the chances of it happening are unlikely.